Require token on most API requests

+++ app/graphql/mutations/delete_favorite_anime.rb
@@ -1,14 +1,17 @@
 module Mutations
   class DeleteFavoriteAnime < Mutations::BaseMutation
-    argument :user_id, ID, required: true
     argument :anime_ids, [ID], required: true
 
     field :deleted, Boolean, null: false
     field :user, Types::UserType
 
-    def resolve(user_id:, anime_ids:)
-      @user = User.find(user_id)
-      @anime = UserAnime.where(user_id: user_id, anime_id: anime_ids)
+    def resolve(anime_ids:)
+      unless context[:ready?]
+        raise GraphQL::ExecutionError.new('Unauthorized Error', options: { status: :unauthorized, code: 401 })
+      end
+
+      @user = context[:current_user]
+      @anime = UserAnime.where(user_id: @user.id, anime_id: anime_ids)
 
       { deleted: @anime.destroy_all, user: @user.reload }
     end
